:sparkles: feat: Add text to speech. Add ConclusionRequest and ArticleRequest models
Added the ConclusionRequest and ArticleRequest models to the infrastructure/web_services/models/conclusion_request directory. These models are used to represent the request data sent to the server when fetching news conclusions.

The ConclusionRequest model contains a list of ArticleRequest objects, representing the articles that will be used to generate the conclusion.

The ArticleRequest model represents an individual article with a title, description, and article text.

These models will be used to serialize the request data before sending it to the server.

This commit implements the necessary changes to add these models to the project.

Signed-off-by: Dmytro Turskyi <dmytro.turskyi@gmail.com>