Seven (#4)
* Introduce devtools options configuration

Adds a `devtools_options.yaml` file for configuring DevTools extension enablement states. This provides users with more granular control over which extensions are enabled or disabled.

Also introduces a new `EndDrawer` widget and a `MarkdownPreview` widget for the user interface.

* :sparkles: implement conclusion feature

Add a conclusion feature that displays a generated summary of the news articles on the home screen.

This feature provides a concise overview of the news, allowing users to quickly grasp the main points without reading all the articles.

The conclusion is generated by sending a request to a backend service with the news articles as input. The service uses a natural language processing model to generate a coherent summary.

- Updated the UI to include a new section for the conclusion.
- Added new API endpoints for generating conclusions.
- Added new state management logic to handle the conclusion state.

* :recycle: refactor: update default country code to USA and fix inconsistency with constants

This commit updates the default country code in the NewsRepositoryImpl, Constants file, NewsRepository interface, RestClient, RestClient generated file, and NewsBloc to 'USA'. It also fixes the inconsistency in the NewsBloc where the default country code was being set based on the condition.

Signed-off-by: Dmytro Turskyi <dmytro.turskyi@gmail.com>

---------

Signed-off-by: Dmytro Turskyi <dmytro.turskyi@gmail.com>